feat(todel): isolate module logic into logic module
